Reworked shootpos, multiple bone setup, switched to threading provided by valve, fixed some animations and velocity, fixed multiple crashes, updated aimbot points
Shootpos is now iterated while applying angles for high precision.
Multiple bone setups for resolving are now done at resolve time to save on preformance.
Switched to valves threading pool and mutexes to reduce crashing.
Fixed some animation issues related to velocity, jumping, and crouching.
Updated aimbot points to provide more accuracy.
Updated resolver shot handling to improve timing/precision.

+++ solace-csgo/address.h
@@ -90,7 +90,7 @@ public:
 		auto out = m_base + offset;
 
 		// get rel32 offset.
-		const auto r = *reinterpret_cast< uintptr_t * >(out);
+		const auto r = *reinterpret_cast< int32_t* >(out);
 		if ( !r )
 			return t{};
